+++ src/main/java/mod/traister101/sns/util/items/ItemContainerSlot.java
@@ -6,12 +6,15 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.items.ItemHandlerHelper;
 
 import lombok.*;
+import lombok.experimental.Accessors;
 import org.jetbrains.annotations.Range;
 
 @ToString
 @AllArgsConstructor(access = AccessLevel.PACKAGE)
 public final class ItemContainerSlot implements ItemSlot {
 
+	@Getter
+	@Accessors(fluent = true)
 	@Range(from = 0, to = Integer.MAX_VALUE)
 	final int slotIndex;
 	final Container container;