+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Fixed
 
-* Attempting to open curio slots using the open hovered slots keybind not working
+*Attempting to open curio slots using the open hovered slots keybind not working
+
+### New
+
+* Added a "priority" config for which curio slot should be prioritized by the Open Item Container keybind